Reorder methods
