+++ psdm_analysis/models/grid_with_results.py
@@ -1,5 +1,4 @@
 import concurrent
-import logging
 from concurrent.futures import ProcessPoolExecutor
 from dataclasses import dataclass
 from datetime import datetime
@@ -10,12 +9,11 @@ from psdm_analysis.models.input.container.grid_container import GridContainer
 from psdm_analysis.models.input.container.participants_container import (
     SystemParticipantsContainer,
 )
-from psdm_analysis.models.input.enums import RawGridElementsEnum, SystemParticipantsEnum
+from psdm_analysis.models.input.enums import RawGridElementsEnum
 from psdm_analysis.models.result.grid.connector import ConnectorsResult
 from psdm_analysis.models.result.grid.enhanced_node import EnhancedNodesResult
 from psdm_analysis.models.result.grid.node import NodesResult
 from psdm_analysis.models.result.grid.transformer import Transformers2WResult
-from psdm_analysis.models.result.participant.participant import ParticipantsResult
 from psdm_analysis.models.result.participant.participants_res_container import (
     ParticipantsResultContainer,
 )
@@ -26,6 +24,78 @@ from psdm_analysis.models.result.res_container import ResultContainer
 class GridWithResults:
     grid: GridContainer
     results: ResultContainer
+
+    def nodal_energies(self) -> dict[str, float]:
+        return {uuid: self.nodal_energy(uuid) for uuid in self.grid.raw_grid.nodes.uuid}
+
+    def nodal_energy(self, uuid: str) -> float:
+        return self.nodal_result(uuid).participants.sum().energy()
+
+    def nodal_results(self) -> dict[str, ResultContainer]:
+        return {
+            node_uuid: self.nodal_result(node_uuid)
+            for node_uuid in self.grid.node_participants_map.keys()
+        }
+
+    def nodal_result(self, node_uuid: str) -> "ResultContainer":
+        node_participants = self.grid.node_participants_map[node_uuid]
+        participants_uuids = node_participants.uuids()
+        participants = self.results.participants.subset(participants_uuids)
+        return ResultContainer(
+            name=node_uuid,
+            nodes=NodesResult(
+                RawGridElementsEnum.NODE,
+                {node_uuid: self.results.nodes.entities[node_uuid]},
+            ),
+            lines=ConnectorsResult.create_empty(RawGridElementsEnum.LINE),
+            transformers_2w=Transformers2WResult.create_empty(
+                RawGridElementsEnum.TRANSFORMER_2_W
+            ),
+            participants=participants,
+        )
+
+    def em_results(
+        self,
+    ) -> List[Tuple[SystemParticipantsContainer, ParticipantsResultContainer]]:
+        uuid_to_connected_asset = self.grid.participants.ems.uuid_to_connected_assets()
+        return [
+            (
+                self.grid.participants.subset(connected_assets + [em_uuid]),
+                self.results.participants.subset(connected_assets + [em_uuid]),
+            )
+            for (em_uuid, connected_assets) in uuid_to_connected_asset.items()
+        ]
+
+    def build_enhanced_nodes_result(self):
+        nodal_results = self.nodal_results()
+
+        with ProcessPoolExecutor() as executor:
+            futures = {
+                executor.submit(self.calc_pq, uuid, nodal_result)
+                for uuid, nodal_result in nodal_results.items()
+            }
+
+            nodal_pq = {}
+            for future in concurrent.futures.as_completed(futures):
+                uuid, pq = future.result()
+                nodal_pq[uuid] = pq
+
+        return EnhancedNodesResult.from_nodes_result(self.results.nodes, nodal_pq)
+
+    def find_participant_result_pair(self, uuid: str):
+        return self.grid.participants.find_participant(
+            uuid
+        ), self.results.participants.find_participant_result(uuid)
+
+    def filter_by_date_time(self, time: Union[datetime, list[datetime]]):
+        return GridWithResults(
+            self.grid.filter_by_date_time(time), self.results.filter_by_date_time(time)
+        )
+
+    def filter_for_time_interval(self, start: datetime, end: datetime):
+        return GridWithResults(
+            self.grid, self.results.filter_for_time_interval(start, end)
+        )
 
     @classmethod
     def from_csv(
@@ -71,109 +141,7 @@ class GridWithResults:
             else GridWithResults(grid, results)
         )
 
-    def nodal_energies(self) -> dict[str, float]:
-        return {uuid: self.nodal_energy(uuid) for uuid in self.grid.raw_grid.nodes.uuid}
-
-    def nodal_energy(self, uuid: str) -> float:
-        return self.nodal_result(uuid).participants.sum().energy()
-
-    def nodal_results(self) -> dict[str, ResultContainer]:
-        return {
-            node_uuid: self.nodal_result(node_uuid)
-            for node_uuid in self.grid.node_participants_map.keys()
-        }
-
-    def nodal_result(self, node_uuid: str) -> "ResultContainer":
-        node_participants = self.grid.node_participants_map[node_uuid]
-        participants_uuids = node_participants.uuids()
-        participants = self.results.participants.subset(participants_uuids)
-        return ResultContainer(
-            name=node_uuid,
-            nodes=NodesResult(
-                RawGridElementsEnum.NODE,
-                {node_uuid: self.results.nodes.entities[node_uuid]},
-            ),
-            lines=ConnectorsResult.create_empty(RawGridElementsEnum.LINE),
-            transformers_2w=Transformers2WResult.create_empty(
-                RawGridElementsEnum.TRANSFORMER_2_W
-            ),
-            participants=participants,
-        )
-
-    # todo: this is not used so might be deleted
-    @staticmethod
-    def _safe_get_result(
-        participant: SystemParticipantsEnum,
-        participant_uuids: List[str],
-        participant_results: "ParticipantsResult",
-    ) -> ParticipantsResult:
-        # todo: link to corresponding SIMONA issue
-        """
-        Returns the corresponding participant results for a list of uuids.
-        Missing results can happen in rare caseswhere SIMONA does not
-        output single results for a system participant.
-
-        :param participant_uuids: the participant uuids to look for
-        :param participant_results: the corresponding results
-        :return: mapping between uuids and results
-        """
-        res = dict()
-        for participant_uuid in participant_uuids:
-            if participant_uuid in participant_results.entities:
-                res[participant_uuid] = participant_results.entities[participant_uuid]
-            else:
-                logging.debug(
-                    "There is no result for result for participant "
-                    + participant_uuid
-                    + "within the participant results of the"
-                    + str(type(participant_results))
-                )
-        return ParticipantsResult(participant, res)
-
-    def em_results(
-        self,
-    ) -> List[Tuple[SystemParticipantsContainer, ParticipantsResultContainer]]:
-        uuid_to_connected_asset = self.grid.participants.ems.uuid_to_connected_assets()
-        return [
-            (
-                self.grid.participants.subset(connected_assets + [em_uuid]),
-                self.results.participants.subset(connected_assets + [em_uuid]),
-            )
-            for (em_uuid, connected_assets) in uuid_to_connected_asset.items()
-        ]
-
-    def build_enhanced_nodes_result(self):
-        nodal_results = self.nodal_results()
-
-        with ProcessPoolExecutor() as executor:
-            futures = {
-                executor.submit(self.calc_pq, uuid, nodal_result)
-                for uuid, nodal_result in nodal_results.items()
-            }
-
-            nodal_pq = {}
-            for future in concurrent.futures.as_completed(futures):
-                uuid, pq = future.result()
-                nodal_pq[uuid] = pq
-
-        return EnhancedNodesResult.from_nodes_result(self.results.nodes, nodal_pq)
-
     @staticmethod
     def calc_pq(uuid, nodal_result: ResultContainer):
         pq = nodal_result.participants.sum()
         return uuid, pq
-
-    def find_participant_result_pair(self, uuid: str):
-        return self.grid.participants.find_participant(
-            uuid
-        ), self.results.participants.find_participant_result(uuid)
-
-    def filter_by_date_time(self, time: Union[datetime, list[datetime]]):
-        return GridWithResults(
-            self.grid.filter_by_date_time(time), self.results.filter_by_date_time(time)
-        )
-
-    def filter_for_time_interval(self, start: datetime, end: datetime):
-        return GridWithResults(
-            self.grid, self.results.filter_for_time_interval(start, end)
-        )
