+++ psdm_analysis/models/input/connector/switches.py
@@ -6,14 +6,6 @@ from psdm_analysis.models.input.enums import EntitiesEnum, RawGridElementsEnum
 
 @dataclass(frozen=True)
 class Switches(Connector):
-    @staticmethod
-    def attributes() -> list[str]:
-        return Connector.attributes() + ["closed"]
-
-    @staticmethod
-    def get_enum() -> EntitiesEnum:
-        return RawGridElementsEnum.SWITCH
-
     @property
     def closed(self):
         return self.data["closed"]
@@ -23,3 +15,11 @@ class Switches(Connector):
 
     def get_opened(self) -> "Switches":
         return self.subset(self.data.query("closed == False").index)
+
+    @staticmethod
+    def attributes() -> list[str]:
+        return Connector.attributes() + ["closed"]
+
+    @staticmethod
+    def get_enum() -> EntitiesEnum:
+        return RawGridElementsEnum.SWITCH
