+++ psdm_analysis/models/input/node.py
@@ -3,30 +3,13 @@ from typing import List
 
 from pandas import DataFrame
 
-from psdm_analysis.models.entity import Entities
+from psdm_analysis.models.input.entity import Entities
 from psdm_analysis.models.input.enums import RawGridElementsEnum
 
 
 @dataclass(frozen=True)
 class Nodes(Entities):
     data: DataFrame
-
-    @staticmethod
-    def get_enum() -> RawGridElementsEnum:
-        return RawGridElementsEnum.NODE
-
-    @staticmethod
-    def attributes() -> List[str]:
-        return Entities.attributes() + [
-            "v_rated",
-            "v_target",
-            "slack",
-            "geo_position",
-            "longitude",
-            "latitude",
-            "volt_lvl",
-            "subnet",
-        ]
 
     @property
     def geo_position(self):
@@ -65,3 +48,20 @@ class Nodes(Entities):
 
     def node(self):
         return self.data.index
+
+    @staticmethod
+    def get_enum() -> RawGridElementsEnum:
+        return RawGridElementsEnum.NODE
+
+    @staticmethod
+    def attributes() -> List[str]:
+        return Entities.attributes() + [
+            "v_rated",
+            "v_target",
+            "slack",
+            "geo_position",
+            "longitude",
+            "latitude",
+            "volt_lvl",
+            "subnet",
+        ]
